Add default query representations for dates and UUIDs (#37)
* Add default query representations for dates and UUIDs

SQLite does not have date or UUID types, and instead can represent dates
and UUIDs in several different ways.

While this works for SQLite, other database systems _do_ have dedicated
date and UUID types, and so we should probably encode these types in
StructuredQueries' decoding and binding layers, and then SQLite drivers
(like SharingGRDB) will pick a sensible default, like ISO8601 strings
for dates, and lowercased strings for UUIDs.

Draft PR for now as we figure out if this is the right direction, and
there is plenty to do before merging (documentation, etc.).

* wip

* wip

* Use StructuredQueriesCore module name in macro expansions. (#42)

* wip

* wip;

* Update docs

* wip

* wip

---------

Co-authored-by: Brandon Williams <mbrandonw@hey.com>